+++ src/api/products/index.ts
@@ -1,28 +1,6 @@
 import { GetProductsParamType } from '@/types';
 import { api } from '@/api/api';
-
-export interface Category {
-  id: number;
-  name: string;
-}
-
-export interface Seller {
-  id: string;
-  nickname: string;
-}
-
-export enum ProductCondition {
-  NEW = 'NEW',
-  SLIGHTLY_USED = 'SLIGHTLY_USED',
-  HEAVILY_USED = 'HEAVILY_USED',
-}
-
-export enum ProductStatus {
-  AVAILABLE = '구매 가능',
-  SOLD_OUT = '판매 완료',
-  RESERVED = '예약중',
-}
-
+import { ProductStatus, ProductCondition, Category, Seller } from '@/types';
 export interface Product {
   id: number;
   title: string;
@@ -48,11 +26,7 @@ export const getCuration = async () => {
   return response.data;
 };
 export const getProduct = async (productId: string): Promise<Product> => {
-  const response = await api.get(`/products/${productId}`, {
-    // headers: {
-    //   Authorization: `Bearer ${AccessToken}`,
-    // }, //아직 로그인과 연결안됨
-  });
+  const response = await api.get(`/products/${productId}`);
   const productData = response.data;
   if (productData.status === 'AVAILABLE') {
     productData.status = '예약중';
